requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
jinja2==3.1.3
python-multipart==0.0.9
itsdangerous==2.1.2
cryptography==42.0.5
requests==2.31.0
pytz==2024.1
pycryptodome==3.20.0
